Hotfix to tolerate server branch limits of 100 and 500 (#3039)
Hooked StreamGetBranches for retry if 500 limit fails

We are catching the SpeckleGraphQLException<StreamData> exception as a way to see that the call to get up to 500 stream branches failed. We can then retry with 100.

While the exception isn't as specific as we would like, it is probably good enough and we don't have anything else to go on, based on how the GQL library is constructed.